Use SAX instead of the ElementTree
We just need to read and parse the file and create our own Python objects.
There's no need to construct the ElementTree (DOM tree) and go through it
afterwards to construct our objects.

This reduces the memory consumption roughly to a half at the cost of prolonged
"start" (import) time.

Also this way we don't need the python-lxml module.

Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>