add entries for several layouts known to be non-ASCII by systemd/s-c-k
systemd contains a store of old Fedora keyboard layout knowledge inherited
from system-config-keyboard. It knows of several layouts that are not
capable of ASCII input, which are not currently listed in langtable. This
patch adds all those layouts to langtable so we preserve the information that
these layouts are not capable of inputting ASCII.