Increase the rank of zh_SG.UTF-8 back to 10 again for languageId="zh"
Contrary to cmn_TW.UTF-8, it is actually useful to offer the zh_SG.UTF-8
locale in the Anaconda language selection screen. zh_SG has no translations
either, nevertheless it is useful to select that locale to get a simplified
Chinese input method by default and the correct time and date formats for
Singapore.